requirements.txt
boto3==1.26.114
openai==0.27.4
openai_whisper==20230314
pydub==0.25.1
SpeechRecognition==3.10.0
PyAudio==0.2.13
python-dotenv
pyttsx3
langchain
wikipedia
wolframalpha
pyowm
pygame
torchaudio
SoundFile
bark @ git+https://github.com/suno-ai/bark@main